CharacterManager changes, refactorings and code cleanup.
This commit renames the CharacterManager to the more appropriate name, AuthProvider, which is functionally the same as the old CharacterManager but refactored. The new AuthProvider exposes and Bindable ObservableCollection of characters to bind to, however this is no longer a global SelectedCharacter that classes can bind to. Instead, tabs should instead create a ComboBox and bind to the ObservableCollection of characters, and use their own selected characters. There are also lots of other refactors in this commit, along with modifications to how the application starts up and what methods are called, including support for async methods to be called before startup is completed.